requirements.txt
schedule~=1.2.0
pandas~=1.4.2
tabulate~=0.8.9
psycopg2~=2.9.6
yfinance~=0.2.18
matplotlib~=3.5.1
seaborn~=0.11.2
pyTelegramBotAPI~=4.11.0
python-dotenv~=1.0.0